change data type of single track vehicle

+++ src/main/java/nex/vts/backend/models/responses/TrackNowResponseTwo.java
@@ -16,7 +16,7 @@ public class TrackNowResponseTwo {
     private Double longs;
 
     @JsonProperty("speed")
-    private Double speed;
+    private Float speed;
 
     @JsonProperty("head")
     private String head;
